Fixed back to front position of splash message
Moved code for splash message style and contents to end of index. This way, the relative position of the box can be set to -1 (very front). I assume that somewhere in the css files the rest of the text automatically goes to very front as well, so by giving the message the last say we make it go to the front